backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6

# Database and caching
SQLAlchemy==2.0.23
redis==5.0.1
Flask-Caching==2.1.0

# Async processing
celery==5.3.4
redis==5.0.1

# Security and validation
bcrypt==4.0.1
WTForms==3.0.1
email-validator==2.0.0
cryptography==41.0.7

# API integrations
requests==2.31.0
httpx==0.25.2
stripe==7.4.0
google-auth==2.23.4
firebase-admin==6.2.0

# Data processing and ML
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1

# Monitoring and logging
sentry-sdk[flask]==1.38.0
structlog==23.2.0
python-json-logger==2.0.7

# File handling
python-multipart==0.0.6
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Environment management
python-dotenv==1.0.0
dynaconf==3.2.4